Extract view model
